requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.109.1
h11==0.14.0
idna==3.7
Logbook==1.5.3
pydantic~=2.9.2
PyJWT==2.7.0
PyMySQL==1.0.3
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.35.0
typing_extensions~=4.8.0
uvicorn==0.22.0
fastapi-healthcheck~=0.2.13
captcha~=0.6.0
aioredis==1.3.1
colorlog
python-dotenv
openpyxl
reportlab